+++ tests/fixtures/self_assert/assertDictEqual_in.py
@@ -9,3 +9,15 @@ class TestDictEqual(TestCase):
 
     def test_simple_msg2(self):
         self.assertDictEqual(klm, 100, "This is wrong!")
+
+    def test_line_wrapping(self):
+        self.assertDictEqual(
+            {
+                'a': 1,
+                'b': 2,
+            },
+            {'b': 2},
+            "This is wrong!",
+        )
+
+        self.assertDictEqual(100, klm)