+++ tests/fixtures/self_assert/assertEqual_out.py
@@ -14,3 +14,20 @@ class TestAssertEqual(TestCase):
     def test_message(self):
         assert 123+z == xxx+y, 'This is wrong!'
         assert 123 == xxx+y, 'This is wrong!'
+
+    def test_line_wrapping(self):
+        assert True == False, 'This will fail %s' % \
+                'always'
+
+        assert 'abc' \
+                         .replace(
+                                  'abc'
+                                  , 'def') == \
+                         'def', \
+                         'Wrap %s' % \
+                         'everything'
+
+    def test_expression_as_argument(self):
+        assert (abc not in self.data) == True
+        assert (abc in self.data) == (not contains)
+        assert contains == (not contains)