+++ tests/fixtures/self_assert/assertRaisesRegex_in.py
@@ -10,5 +10,13 @@ class TestRaises(TestCase):
     def test_kwargs(self):
         self.assertRaisesRegex(RunTimeError, pattern, someFunc, foo=42, bar=43)
 
+    def test_context_manager(self):
+        with self.assertRaisesRegex(RunTimeError, pattern):
+            someFunc('foo', None)
+
     def test_args_kwargs(self):
         self.assertRaisesRegex(RunTimeError, pattern, someFunc, 1,2,3, foo=42, bar=43)
+
+    def test_lambda(self):
+        self.assertRaises(RunTimeError, lambda: error(1, 2))
+        self.assertRaises(RunTimeError, lambda: error(1, 2) or error())