+++ tests/fixtures/self_assert/assertRaisesRegexp_out.py
@@ -1,5 +1,7 @@
 # required-method: assertRaisesRegexp
 
+import pytest
+import re
 class TestRaises(TestCase):
     def test_simple(self):
         with pytest.raises(RunTimeError) as excinfo:
@@ -16,7 +18,18 @@ class TestRaises(TestCase):
             someFunc(foo=42, bar=43)
         assert re.search(pattern, excinfo.value)
 
+    def test_context_manager(self):
+        with pytest.raises(RunTimeError) as excinfo:
+            someFunc('foo', None)
+        assert re.search(pattern, excinfo.value)
+
     def test_args_kwargs(self):
         with pytest.raises(RunTimeError) as excinfo:
             someFunc(1,2,3, foo=42, bar=43)
         assert re.search(pattern, excinfo.value)
+
+    def test_lambda(self):
+        with pytest.raises(RunTimeError):
+            error(1, 2)
+        with pytest.raises(RunTimeError):
+            error(1, 2) or error()