+++ tests/fixtures/self_assert/assertRaises_out.py
@@ -1,7 +1,17 @@
 # required-method: assertRaises
+import os
+from os import listdir, path
+import subprocess as sb
+import pytest
+
+class Dummy():
+    pass
 
 class TestRaises(TestCase):
     def test_simple(self):
+        import io
+        from io import StringIO, BytesIO
+        import numpy as np
         with pytest.raises(RunTimeError):
             someFunc()
 
@@ -16,3 +26,18 @@ class TestRaises(TestCase):
     def test_args_kwargs(self):
         with pytest.raises(RunTimeError):
             someFunc(1,2,3, foo=42, bar=43)
+
+    def test_context_manager(self):
+        with pytest.raises(RunTimeError):
+            someFunc()
+
+    def test_context_manager_var(self):
+        with pytest.raises(RunTimeError) as ctx:
+            someFunc()
+        assert ctx.exception
+
+    def test_lambda(self):
+        with pytest.raises(RunTimeError):
+            error(1, 2)
+        with pytest.raises(RunTimeError):
+            error(1, 2) or error()