+++ tests/fixtures/self_assert/assertRegexpMatches_out.py
@@ -1,5 +1,6 @@
 # required-method: assertRegexpMatches
 
+import re
 class TestAssertEqual(TestCase):
     def test_you(self):
         assert re.search('xxx', abc)