+++ tests/fixtures/self_assert/assertTupleEqual_out.py
@@ -9,3 +9,13 @@ class TestTupleEqual(TestCase):
 
     def test_simple_msg2(self):
         assert klm == 100, "This is wrong!"
+
+    def test_line_wrapping(self):
+        assert (
+                'a',
+                'b'
+            ) == \
+            ('b',), \
+            "This is wrong!"
+
+        assert 100 == klm