+++ CHANGES.rst
@@ -1,11 +1,24 @@
 Changelog
 ============
 
+v0.3 (2016-07-26)
+----------------
+
+* Add support for assertRaises / assertWarns context managers.
+
+* Add support for converting lambda arguments in assertRaises into
+  context managers.
+
+* Fix some incorrect transformations.
+
+* Internal cleanup and fixes.
+
+
 v0.2 (2015-10-20)
 ---------------------
 
-* Add support assertRegex/assertRegexpMatches, assertNotRegex
-  assertRaisesRegex/assertRaisesRegexp assertWarnsRegex.
+* Add support for assertRegex/assertRegexpMatches, assertNotRegex,
+  assertRaisesRegex/assertRaisesRegexp, assertWarnsRegex.
 
 * `unittest2pytest` is now a `pytest` subproject.